requirements.txt
fastapi==0.112.2
pydantic-settings==2.4.0
#psycopg2==2.9.9
psycopg2-binary==2.9.9
redis==5.0.8
pycryptodome==3.20.0
python-jose==3.3.0
celery==5.4.0
python-dotenv==1.0.1